+++ src/capstone2llvmir/x86/x86_impl.h
@@ -136,6 +136,7 @@ class Capstone2LlvmIrTranslatorX86_impl :
 				llvm::IRBuilder<>& irb,
 				llvm::AtomicRMWInst::BinOp aop);
 		bool tryTranslateLockedCmpxchg(cs_insn* i, cs_x86* xi, llvm::IRBuilder<>& irb);
+		bool tryTranslateLockedCmpxchgWide(cs_insn* i, cs_x86* xi, llvm::IRBuilder<>& irb, unsigned bits);
 		bool tryTranslateLockedIncDec(cs_insn* i, cs_x86* xi, llvm::IRBuilder<>& irb, bool isDec);
 
 		unsigned getAddrSpace(x86_reg segment);