+++ tests/capstone2llvmir/arm64_tests.cpp
@@ -3778,6 +3778,25 @@ TEST_P(Capstone2LlvmIrTranslatorArm64Tests, StrStoreIsNotAtomic)
 	}
 }
 
+TEST_P(Capstone2LlvmIrTranslatorArm64Tests, StxpStoreIsAtomic)
+{
+	auto* f = translate(assemble("stxp w0, w1, w2, [x3]"));
+	ASSERT_NE(nullptr, f);
+	bool found = false;
+	for (auto it = inst_begin(f), e = inst_end(f); it != e; ++it)
+	{
+		if (auto* s = dyn_cast<StoreInst>(&*it))
+		{
+			if (s->isAtomic())
+			{
+				found = true;
+				break;
+			}
+		}
+	}
+	EXPECT_TRUE(found);
+}
+
 //
 // ARM64_INS_LDXRB
 //